+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+* Added a `bindPath` option to the NixOS module.
+
+### Removed
+* Removed the `bind_addrs` configuration option.
+
 ## [1.1.0] - 2024-07-20
 ### Added
 * Added pretty page titles. Pages now display the HTML title instead of directory path.
@@ -20,6 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[unreleased]: https://github.com/newAM/oidc_pages/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/newAM/oidc_pages/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/newAM/oidc_pages/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/newAM/oidc_pages/releases/tag/v1.0.0